- ensured header files' guards start with LDFORGE_ - removed typedef "qchar" (use QChar instead), removed use of short, long and size_t (use int instead) - use C++11-using syntax instead of typedefs in types.h

+++ src/addObjectDialog.h
@@ -59,10 +59,10 @@ class AddObjectDialog : public QDialog
 		QLineEdit* le_matrix;
 
 	private:
-		void setButtonBackground (QPushButton* button, short color);
+		void setButtonBackground (QPushButton* button, int color);
 		str currentSubfileName();
 
-		short colnum;
+		int colnum;
 
 	private slots:
 		void slot_colorButtonClicked();
